Preserve fractional letter spacing in the Flutter editor (#2)
* feat!: preserve fractional letter spacing in editor fields

* test: cover fractional letter spacing editing and history

* chore(release): prepare canvas_editor_flutter 0.2.0

+++ packages/canvas_editor_flutter/test/asset_library_extension_test.dart
@@ -26,7 +26,7 @@ final class _FakeTextMeasurer implements TextMeasurer {
     required String fontFamily,
     required int fontWeight,
     required double fontSize,
-    required int letterSpacing,
+    required double letterSpacing,
   }) {
     return Size2D(text.length * fontSize * 0.6, fontSize);
   }
