+++ packages/canvas_editor_flutter/test/editor_export_scene_boundary_test.dart
@@ -20,7 +20,7 @@ class _FakeTextMeasurer implements TextMeasurer {
     required String fontFamily,
     required int fontWeight,
     required double fontSize,
-    required int letterSpacing,
+    required double letterSpacing,
   }) {
     return Size2D(text.length * fontSize * 0.6, fontSize);
   }
